requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
openpyxl==3.1.5
APScheduler==3.10.4
python-dateutil==2.9.0.post0
